Refactor Code Project
