+++ src/components/atoms/input/input.tsx
@@ -1,6 +1,5 @@
 import * as React from "react"
-
-import { cn } from "../../lib/utils"
+import { cn } from "../../../lib/utils"
 
 const Input = React.forwardRef<HTMLInputElement, React.ComponentProps<"input">>(
   ({ className, type, ...props }, ref) => {
