+++ src/components/atoms/tooltip/tooltip.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as TooltipPrimitive from "@radix-ui/react-tooltip"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../../lib/utils"
 
 const TooltipProvider = TooltipPrimitive.Provider
