+++ src/components/molecules/home/HomeBody.tsx
@@ -1,7 +1,4 @@
-import React from "react";
 import styled from "styled-components";
-
-
 
 type Props = {};
 export const HomeBody = (props: Props) => {
